requirements.txt
scikit-learn
torch
transformers
sentence-transformers
opencv-python
Pillow
python-dotenv
google-generativeai
gtts
pydub
pygame